+++ src/main/kotlin/com/screentech/cordamigrate/dao/securityquestion/SecurityQuestionRepository.kt
@@ -4,6 +4,6 @@ import com.screentech.cordamigrate.entity.securityquestion.SecurityQuestion
 import org.springframework.data.jpa.repository.JpaRepository
 import java.math.BigDecimal
 
-interface SecurityQuestionRepository : JpaRepository<SecurityQuestion, BigDecimal> {
+interface SecurityQuestionRepository : JpaRepository<SecurityQuestion, Long> {
 
 }
