+++ src/main/kotlin/com/screentech/cordamigrate/dao/securityquestion/UserAnswerRepository.kt
@@ -4,6 +4,6 @@ import com.screentech.cordamigrate.entity.securityquestion.UserAnswer
 import org.springframework.data.jpa.repository.JpaRepository
 import java.math.BigDecimal
 
-interface UserAnswerRepository : JpaRepository<UserAnswer, BigDecimal> {
+interface UserAnswerRepository : JpaRepository<UserAnswer, Long> {
 
 }
