+++ src/main/kotlin/com/screentech/cordamigrate/utility/CRUDInterface.kt
@@ -14,11 +14,11 @@ interface CRUDInterface <T> {
 
     fun findAll() :  ResponseEntity<*>
 
-    fun findById(id: BigDecimal) : ResponseEntity<*>
+    fun findById(id: Long) : ResponseEntity<*>
 
     fun findByEmail(email : String) : ResponseEntity<*>
 
-    fun deleteById(id: BigDecimal) : ResponseEntity<*>
+    fun deleteById(id: Long) : ResponseEntity<*>
 
     fun deleteByObject(anObject: T) : ResponseEntity<*>
 }
